Linux 64bit changes
- Some of these changes need checked and over looked and what not to work better with other platforms as well. but apart from that this works linux 64bit though this should be updated to sdl2.0

+++ engine/source/platform/types.posix.h
@@ -27,9 +27,10 @@
 #define FN_CDECL     ///< Calling convention
 
 // size_t is needed to overload new
-// size_t tends to be OS and compiler specific and may need to 
+// size_t tends to be OS and compiler specific and may need to
 // be if/def'ed in the future
-typedef unsigned int  dsize_t;      
+#include <stddef.h>
+typedef size_t   dsize_t;
 
 
 /** Platform dependent file date-time structure.  The defination of this structure
